Fix memory allocator bugs

+++ kernel/modules/lib/bit.v
@@ -8,7 +8,7 @@ pub fn bittest(bitmap Bitmap, index u64) bool {
 		bits_type := sizeof(u64) * 8
 		test_index := index % bits_type
 		test_sample := fbitmap[index / bits_type]
-		return ((test_sample >> test_index) & 1) != 0
+		return ((test_sample >> test_index) & u64(1)) != 0
 	}
 }
 
@@ -17,7 +17,7 @@ pub fn bitset(bitmap Bitmap, index u64) {
 		mut fbitmap := &u64(bitmap)
 		bits_type := sizeof(u64) * 8
 		test_index := index % bits_type
-		fbitmap[index / bits_type] |= 1 << test_index
+		fbitmap[index / bits_type] |= u64(1) << test_index
 	}
 }
 
@@ -26,6 +26,6 @@ pub fn bitreset(bitmap Bitmap, index u64) {
 		mut fbitmap := &u64(bitmap)
 		bits_type := sizeof(u64) * 8
 		test_index := index % bits_type
-		fbitmap[index / bits_type] &= ~(1 << test_index)
+		fbitmap[index / bits_type] &= ~(u64(1) << test_index)
 	}
 }
